Setting up filter by difficulty functionality
