+++ src/Components/Views/SearchChallenges/DifficultyLevels.js
@@ -1,10 +1,11 @@
 import React, { useState } from 'react';
+import jssPluginPropsSort from 'jss-plugin-props-sort';
 
-const DifficultyLevels = () => {
+const DifficultyLevels = (props) => {
 
   return (
     <div >
-        <select>
+        <select onChange={(e) => props.filterByDifficulty(e.target.value)}>
             <option value="easy">Easy</option>
             <option value="medium">Medium</option>
             <option value="hard">Hard</option>
